Admin Module (Added fragments in master module html)

+++ src/main/java/com/erp/mastro/controller/UserController.java
@@ -1,8 +1,9 @@
 package com.erp.mastro.controller;
 
-//import com.erp.mastro.common.MailUtils;
 import com.erp.mastro.common.MailUtils;
+import com.erp.mastro.common.MastroLogUtils;
 import com.erp.mastro.custom.responseBody.GenericResponse;
+import com.erp.mastro.dto.CurrentUserDetails;
 import com.erp.mastro.entities.*;
 import com.erp.mastro.model.request.UserModel;
 import com.erp.mastro.service.interfaces.BranchService;
@@ -10,6 +11,8 @@ import com.erp.mastro.service.interfaces.EmployeeService;
 import com.erp.mastro.service.interfaces.RolesService;
 import com.erp.mastro.service.interfaces.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -41,31 +44,84 @@ public class UserController {
     @Autowired
     EmployeeService employeeService;
 
+    /**
+     * Method to login
+     *
+     * @param model
+     * @return to the login page
+     */
     @RequestMapping("/")
     public String welcome(Map<String, Object> model) {
         return "views/login";
     }
 
+    /**
+     * Method to login
+     *
+     * @param model
+     * @return to the login page
+     */
+    @RequestMapping("/login")
+    public String login(Map<String, Object> model) { return "views/login"; }
+
+    /**
+     * Method to view dashboard
+     *
+     * @param model
+     * @return selected branch and branch list
+     */
     @RequestMapping("/home")
-    public String home(Map<String, Object> model) {
+    public String home(Model model) {
+        User userDetails = getCurrentUser();
+        List<Branch> branchList = new ArrayList<>();
+        String currentBranch = null;
+        if(null != userDetails) {
+            for (Branch branch : userDetails.getBranch()) {
+                branchList.add(branch);
+            }
+            if(userDetails.getUserSelectedBranch() != null && userDetails.getUserSelectedBranch().getCurrentBranch() != null) {
+                currentBranch = userDetails.getUserSelectedBranch().getCurrentBranch().getBranchName();
+            }
+        }
+        model.addAttribute("currentBranch",currentBranch);
+        model.addAttribute("branchList",branchList);
         return "views/dashboard";
     }
 
-    @RequestMapping("/login")
-    public String login(Map<String, Object> model) {
-        return "views/login";
+    /**
+     * Method to get the selected branch
+     *
+     * @param branchId
+     * @param model
+     * @return selected branch name
+     */
+    @PostMapping("/admin/selectBranchName")
+    public String getBranchName( Model model, @Valid Long branchId) {
+        Branch branch = branchService.getBranchById(branchId);
+        String branchName = branch.getBranchName();
+        User userDetails = getCurrentUser();
+        userService.saveCurrentBranch(branchId,userDetails);
+        model.addAttribute("branchName",branchName);
+        return "redirect:/home";
     }
 
+    /**
+     * Method to redirect to access denied
+     *
+     * @return to the access-denied page
+     */
     @GetMapping("/access-denied")
-    public String accessDenied() {
-        return "/error/access-denied";
-    }
+    public String accessDenied() { return "/error/access-denied"; }
 
+    /**
+     * Method to add new User
+     *
+     * @param model
+     * @return added user list with branch and role list
+     */
     @RequestMapping("/admin/addUser")
     public String addUser(Model model) {
-
-
-
+        MastroLogUtils.info(UserController.class, "Going to add User :{}");
         try {
            /*List<User> userList = new ArrayList<>();
             userList = userService.getAllUsers();*/
@@ -75,8 +131,7 @@ public class UserController {
                 }*//*
               *//*  if (user.isEnabled()) {
                     userList.add(user);
-                }
-*//*
+                }*//*
             }*/
 
             List<Roles> rolesList = new ArrayList<>();
@@ -105,16 +160,21 @@ public class UserController {
             model.addAttribute("branchList",branchList);
             return "views/user_master";
         } catch (Exception e) {
+            MastroLogUtils.error(UserController.class, "Error occured while adding user : {}");
             throw e;
         }
     }
 
-
+    /**
+     * Method to save User for test
+     *
+     * @return saved user
+     */
     @PostMapping(value = "/register")
     public void register() {
 
         User user = new User();
-        user.setUserName("Ranjit");
+        user.setUserName("ranjit@halo.ae");
         user.setEmail("ranjit@halo.ae");
         user.setPassword(bCryptPasswordEncoder.encode("ranjit"));
         user.setEnabled(true);
@@ -128,14 +188,32 @@ public class UserController {
         userService.savUser(user);
     }
 
+    /**
+     * Method to register User
+     *
+     * @param model
+     * @param request
+     * @param userModel
+     * @return saved User details
+     */
     @PostMapping(value = "/admin/registerUser")
     public String register(@ModelAttribute ("addUserForm") @Valid UserModel userModel, HttpServletRequest request, Model model) {
+        MastroLogUtils.info(UserController.class, "Going to save User :{}");
+        try {
+            userService.saveOrUpdateUser(userModel,request);
+            return "redirect:/admin/addUser";
+        } catch (Exception e) {
+            MastroLogUtils.error(UserController.class, "Error occured while registering user : {}", e);
+            throw e;
+        }
 
-        userService.saveOrUpdateUser(userModel,request);
-
-        return "redirect:/admin/addUser";
     }
 
+    /**
+     * Method for autocomplete employee email address
+     *
+     * @return email address of te employee
+     */
     @RequestMapping(value = "/admin/employeeAutoComplete")
     @ResponseBody
     public GenericResponse EmployeeAutoComplete() {
@@ -149,11 +227,20 @@ public class UserController {
                 .setProperty("value", emails);
     }
 
+    /**
+     * Method to edit User
+     *
+     * @param model
+     * @param request
+     * @param userId
+     * @return edited User details
+     */
     @GetMapping("/admin/getUserForEdit")
     @ResponseBody
     public GenericResponse getRoleForEdit(Model model, HttpServletRequest request, @RequestParam("userId") Long userId) {
-         User userDetails = userService.getUserById(userId);
 
+        MastroLogUtils.info(UserController.class, "Going to edit User :{}" + userId);
+        User userDetails = userService.getUserById(userId);
         Set<UserModel.UserModelEdit> userModelEdits = new HashSet<>();
          for (Roles roles : userDetails.getRoles() ){
                 UserModel.UserModelEdit userEditModel = new  UserModel.UserModelEdit();
@@ -167,17 +254,14 @@ public class UserController {
                 rolesList.add(roles);
             }
         }
-        //full roles
         Set<UserModel.UserModelEdit> rolemodelEdits = new HashSet<>();
-       for (Roles roles : rolesList ){
-           UserModel.UserModelEdit rolemodelEdit = new  UserModel.UserModelEdit();
-           rolemodelEdit.setRole(roles.getRoleName());
-           rolemodelEdit.setId(roles.getId());
-           rolemodelEdits.add(rolemodelEdit);
+        for (Roles roles : rolesList ){
+            UserModel.UserModelEdit rolemodelEdit = new  UserModel.UserModelEdit();
+            rolemodelEdit.setRole(roles.getRoleName());
+            rolemodelEdit.setId(roles.getId());
+            rolemodelEdits.add(rolemodelEdit);
         }
-
-
-            List<Branch> branchList = new ArrayList<>();
+        List<Branch> branchList = new ArrayList<>();
         for (Branch branch : branchService.getAllBranch()) {
             if (branch.getBranchDeleteStatus() != 1) {
                 branchList.add(branch);
@@ -190,7 +274,6 @@ public class UserController {
             branchmodelEdit.setId(branch.getId());
             branchmodelEdits.add(branchmodelEdit);
         }
-
         List<UserModel.UserModelBranchEdit> userModelBranchEdits = new ArrayList<>();
         for (Branch branch : userDetails.getBranch() ){
             UserModel.UserModelBranchEdit editBranch = new UserModel.UserModelBranchEdit();
@@ -198,7 +281,6 @@ public class UserController {
             editBranch.setId(branch.getId());
             userModelBranchEdits.add(editBranch);
         }
-
         return new GenericResponse(true,"get User details")
                 .setProperty("userId",userDetails.getId())
                 .setProperty("email",userDetails.getEmail())
@@ -208,40 +290,63 @@ public class UserController {
                 .setProperty("branch",userModelBranchEdits);
     }
 
+    /**
+     * Method to Activate or Deactivate User
+     *
+     * @param model
+     * @param request
+     * @param userId
+     * @return activated or deactivated userId
+     */
     @GetMapping("/admin/getActivateOrDeactivateUser")
     @ResponseBody
     public GenericResponse getUserActivate(Model model, HttpServletRequest request, @RequestParam("userId") Long userId) {
-         User userDetails = userService.getUserById(userId);
+        MastroLogUtils.info(UserController.class, "Going to Activate or Deactivate User :{}" + userId);
+        User userDetails = userService.getUserById(userId);
         userService.activateOrDeactivateUser(userId);
-      //  boolean isEnabled = userDetails.isEnabled();
         return new GenericResponse(true,"get User details")
                 .setProperty("userId",userDetails.getId());
-               // .setProperty("status",isEnabled);
     }
 
+    /**
+     * Method to delete User
+     *
+     * @param model
+     * @param request
+     * @param userId
+     * @return deleted User list
+     */
     @PostMapping("/admin/deleteUserDetails")
     @ResponseBody
     public GenericResponse deleteRoles(Model model, HttpServletRequest request, @RequestParam("userId") Long userId) {
 
+        MastroLogUtils.info(UserController.class, "Going to delete User : {}");
         try {
-            userService.deleteUserDetails(userId);
-            return new GenericResponse(true, "delete User Details");
+            if (userId != null) {
+                userService.deleteUserDetails(userId);
+                return new GenericResponse(true, "delete User Details");
+            } else {
+                return new GenericResponse(false, "User id null");
+            }
         } catch (Exception e) {
+            MastroLogUtils.error(UserController.class, "Error occured while deleting User : {}", e);
             return new GenericResponse(false, e.getMessage());
         }
     }
 
-    @RequestMapping("/admin/getUserBranches")
-    public GenericResponse getBranchByUserId(Model model, @RequestParam("userId") Long userId) {
-        System.out.println("Inside getBranchByUserId");
-        User userDetails = userService.getUserById(userId);
-        List<Branch> branchList = new ArrayList<>();
-        for (Branch branch : userDetails.getBranch()) {
-            System.out.println("branch :: " + branch);
-            branchList.add(branch);
+    /**
+     * Method to get current User
+     *
+     * @return current logged in User details
+     */
+    public User getCurrentUser() {
+        User userDetails = null;
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if(auth.isAuthenticated()) {
+            CurrentUserDetails currentUser = (CurrentUserDetails) auth.getPrincipal();
+            userDetails = userService.findByEmail(currentUser.getUser().getEmail());
         }
-        return new GenericResponse(true,"get User details")
-        .setProperty("branchList",branchList);
+        return userDetails;
     }
 
 }
